build(olgis): 增加maven前后端统一构建配置 修改olgis前端app
