+++ olgis-gisservice/src/main/resources/application.yml
@@ -5,6 +5,16 @@ server:
     uri-encoding: utf-8
 
 spring:
+
+  thymeleaf:
+    prefix: classpath:/templates/
+    suffix: .html
+    mode: HTML5
+    encoding: UTF-8
+    servlet:
+      content-type: text/html
+    cache: false
+
   datasource:
     url: jdbc:postgresql://localhost:5432/postgres
     driver-class-name: org.postgresql.Driver
