+++ olgis-web-ts/package.json
@@ -2,6 +2,7 @@
   "name": "olgis-web-ts",
   "version": "0.1.0",
   "private": true,
+  "homepage": ".",
   "dependencies": {
     "@material-ui/core": "^4.10.1",
     "@material-ui/icons": "^4.9.1",
