+++ olgis-web-ts/src/apps/OLGISApp/Panel/panel.stories.tsx
@@ -1,7 +1,7 @@
 import {storiesOf} from "@storybook/react";
 import Panel from "./panel";
 import React from "react";
-import Map from "../Map/map";
+import Map from "../../../components/Map/map";
 
 storiesOf("左侧功能栏", module)
 .add("左侧功能栏", ()=>(
