+++ olgis-web-ts/src/apps/OLGISApp/Panel/panel.style.ts
