+++ olgis-web-ts/src/components/TOC/layerItem.tsx
@@ -5,7 +5,7 @@ import LayerInfoProps from "./layerInfoProps";
 import LayerUpButton from "./layerUpButton";
 import LayerDownButton from "./layerDownButton";
 import LayerDeleteIconButton from "./layerDeteteIconButton";
-import {ListItemButton} from "../Panel/expansionPanel";
+import {ListItemButton} from "../tools/expansionPanel";
 import LayerSettingButton from "./layerSettingButton";
 
 interface LayerItemProps extends LayerInfoProps{
