+++ olgis-web-ts/src/stories/index.tsx
@@ -1,6 +1,6 @@
 import React, {ReactNode} from "react";
 import Map from "../components/Map/map";
-import Panel from "../components/Panel/panel";
+import Panel from "../apps/OLGISApp/Panel/panel";
 
 export const showWithMap = (children:ReactNode) => {
     return () => (
